Add validation, some corrects and finally working update photo in Travel View

+++ lib/screens/travel_view.dart
@@ -67,7 +67,7 @@ class _TravelViewState extends State<TravelView> {
         centerTitle: true,
       ),
       floatingActionButton: Padding(
-        padding: const EdgeInsets.only(bottom: 90.0, right: 20.0),
+        padding: const EdgeInsets.only(bottom: 30.0, right: 10.0),
         child: FloatingActionButton(
           child: Icon(Icons.add),
           backgroundColor: Colors.blue,
@@ -78,10 +78,10 @@ class _TravelViewState extends State<TravelView> {
         ),
       ),
       body: StreamBuilder<QuerySnapshot>(
-    stream: FirebaseFirestore.instance
-        .collection("Travel")
-        .where("userId", isEqualTo: FirebaseAuth.instance.currentUser!.uid)
-        .snapshots(),
+        stream: FirebaseFirestore.instance
+            .collection("Travel")
+            .where("userId", isEqualTo: FirebaseAuth.instance.currentUser!.uid)
+            .snapshots(),
         builder: (context, snapshot) {
           if (snapshot.connectionState == ConnectionState.waiting) {
             return Center(child: CircularProgressIndicator());
@@ -97,38 +97,18 @@ class _TravelViewState extends State<TravelView> {
               return InkWell(
                   onTap: () => navigateToDetail(snapshot.data!.docs[index]),
                   child: Container(
-                    margin: EdgeInsets.symmetric(vertical: 5),
+                    margin: EdgeInsets.symmetric(
+                        vertical: 5, horizontal: 10),
                     child: Card(
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(15),
+                      ),
                       child: Padding(
                         padding: EdgeInsets.all(5),
                         child: Container(
                           padding: EdgeInsets.all(5),
                           child: Row(
                             children: [
-                              // CircleAvatar(
-                              //   radius: 30,
-                              //   backgroundImage: NetworkImage(model.photo as String),
-                              // backgroundImage: NetworkImage((FirebaseStorage.instance.ref().child(model.photo as String).getDownloadURL()).toString())
-                              // ),
-                              // FutureBuilder(
-                              //   future: FirebaseStorage.instance
-                              //       .refFromURL(model.photo as String)
-                              //       .getDownloadURL(),
-                              //   builder: (BuildContext context,
-                              //       AsyncSnapshot<dynamic> snapshot) {
-                              //     if (snapshot.hasData) {
-                              //       gowno = (model.photo as String);
-                              //       return CircleAvatar(
-                              //         radius: 40,
-                              //         backgroundImage: NetworkImage(
-                              //             snapshot.data.toString()),
-                              //       );
-                              //     } else {
-                              //       return Center(
-                              //           child: CircularProgressIndicator());
-                              //     }
-                              //   },
-                              // ),
                               FutureBuilder(
                                 future: FirebaseStorage.instance
                                     .refFromURL(model.photo as String)
@@ -136,7 +116,6 @@ class _TravelViewState extends State<TravelView> {
                                 builder: (BuildContext context,
                                     AsyncSnapshot<dynamic> snapshot) {
                                   if (snapshot.hasData) {
-                                    img = (model.photo as String);
                                     return CircleAvatar(
                                       radius: 50,
                                       backgroundImage:
@@ -175,53 +154,30 @@ class _TravelViewState extends State<TravelView> {
                                 child: Container(),
                               ),
                               PopupMenuButton(
-                                icon:
-                                    Icon(Icons.more_vert, color: Colors.black),
-                                itemBuilder: (BuildContext context) => [
-                                  PopupMenuItem(
-                                    child: Text("Edit"),
-                                    value: "edit",
-                                  ),
-                                  PopupMenuItem(
-                                    child: Text("Delete"),
-                                    value: "delete",
-                                  ),
-                                ],
-                                onSelected: (value) {
-                                  if (value == "edit") {
-                                    navigateToUpdate(snapshot.data!.docs[index],
-                                        snapshot.data!.docs[index].id);
-                                  } else {
-                                    showDialog(
-                                      context: context,
-                                      builder: (BuildContext context) {
-                                        return AlertDialog(
-                                          title: const Text('Delete Travel'),
-                                          content: const Text(
-                                              'Are you sure you want to delete this travel?'),
-                                          actions: [
-                                            TextButton(
-                                              onPressed: () =>
-                                                  Navigator.pop(context),
-                                              child: const Text('No'),
-                                            ),
-                                            TextButton(
-                                              onPressed: () {
-                                                deleteTravel(
-                                                    snapshot.data!.docs[index].id, img);
-                                                Navigator.pop(context);
-                                              },
-                                              child: const Text('Yes'),
-                                            ),
-                                          ],
-                                        );
-                                      },
-                                    );
-                                    // deleteTravel(
-                                    //     snapshot.data!.docs[index].id, gowno);
-                                  }
-                                },
-                              ),
+                                  icon: Icon(Icons.more_vert,
+                                      color: Colors.black),
+                                  itemBuilder: (BuildContext context) => [
+                                        PopupMenuItem(
+                                          child: Text("Edit"),
+                                          value: "edit",
+                                        ),
+                                        PopupMenuItem(
+                                          child: Text("Delete"),
+                                          value: "delete",
+                                        ),
+                                      ],
+                                  onSelected: (value) {
+                                    if (value == "edit") {
+                                      navigateToUpdate(
+                                          snapshot.data!.docs[index],
+                                          snapshot.data!.docs[index].id);
+                                    } else if (value == "delete") {
+                                      deleteTravel(
+                                          snapshot.data!.docs[index].id,
+                                          snapshot.data!.docs[index]["photo"]
+                                              as String);
+                                    }
+                                  }),
                             ],
                           ),
                         ),
@@ -235,19 +191,19 @@ class _TravelViewState extends State<TravelView> {
     );
   }
 
-  void deleteTravel(documentId, String imageId) async {
-    var db = FirebaseFirestore.instance;
-    //dodac if że ten imageId musi być równe tego co jest w bazie by usunac to co jest w ID w bazie firebase
-    FirebaseStorage.instance.refFromURL(imageId).delete().then((_) {
-      print("Image successfully deleted!");
-    }).catchError((error) {
-      print("Error removing image: $error");
-    });
-
-    db.collection("Travel").doc(documentId).delete().then((_) {
-      print("Document successfully deleted!");
-    }).catchError((error) {
-      print("Error removing document: $error");
+  void deleteTravel(String id, String imageId) {
+    FirebaseFirestore.instance
+        .collection("Travel")
+        .doc(id)
+        .get()
+        .then((DocumentSnapshot documentSnapshot) {
+      if (documentSnapshot.exists) {
+        FirebaseStorage.instance.refFromURL(imageId).delete().then((_) {
+        }).catchError((error) {
+          print("Error deleting image: $error");
+        });
+        documentSnapshot.reference.delete();
+      }
     });
   }
 }
